[Refactor] Move page handle and connector predicate parser dependencies
