+++ be/src/compute_env/runtime_range_pruner.hpp
@@ -19,7 +19,9 @@
 #include <utility>
 
 #include "column/global_dict/config.h"
+#include "compute_env/runtime_range_pruner.h"
 #include "exec/runtime_filter/runtime_filter_probe.h"
+#include "runtime/descriptors.h"
 #include "runtime/runtime_in_filter.h"
 #include "storage/primitive/column_and_predicate.h"
 #include "storage/primitive/column_or_predicate.h"
@@ -27,7 +29,6 @@
 #include "storage/primitive/column_value_range.h"
 #include "storage/primitive/filter_condition.h"
 #include "storage/primitive/predicate_parser.h"
-#include "storage/runtime_range_pruner.h"
 
 namespace starrocks {
 namespace detail {
