+++ be/src/exec/hdfs_scanner/hdfs_scanner_context.h
@@ -36,8 +36,8 @@
 #include "formats/scan_context.h"
 #include "fs/fs.h"
 #include "runtime/descriptors.h"
-#include "storage/predicate_parser.h"
 #include "storage/primitive/column_predicate_factory.h"
+#include "storage/primitive/predicate_parser.h"
 #include "storage/primitive/predicate_tree/predicate_tree.h"
 
 namespace starrocks {
