+++ be/src/exec/olap_scan_prepare.cpp
@@ -26,6 +26,8 @@
 #include "common/config_scan_io_fwd.h"
 #include "common/object_pool.h"
 #include "compute_env/global_dict/fragment_dict_state.h"
+#include "compute_env/runtime_range_pruner.h"
+#include "compute_env/runtime_range_pruner.hpp"
 #include "exprs/binary_predicate.h"
 #include "exprs/compound_predicate.h"
 #include "exprs/dictmapping_expr.h"
@@ -43,8 +45,6 @@
 #include "storage/primitive/column_predicate_factory.h"
 #include "storage/primitive/predicate_tree/predicate_tree.hpp"
 #include "storage/runtime_filter_predicate.h"
-#include "storage/runtime_range_pruner.h"
-#include "storage/runtime_range_pruner.hpp"
 #include "types/date_value.h"
 #include "types/logical_type.h"
 #include "types/logical_type_infra.h"
