+++ be/src/exec/pipeline/scan/olap_chunk_source.cpp
@@ -31,6 +31,7 @@
 #include "common/statusor.h"
 #include "common/util/table_metrics.h"
 #include "compute_env/global_dict/fragment_dict_state.h"
+#include "compute_env/runtime_range_pruner.hpp"
 #include "compute_env/workgroup/work_group.h"
 #include "exec/catalog_scan_metrics.h"
 #include "exec/olap_scan_node.h"
@@ -62,7 +63,6 @@
 #include "storage/predicate_parser.h"
 #include "storage/primitive/projection_iterator.h"
 #include "storage/primitive/vector_search_option.h"
-#include "storage/runtime_range_pruner.hpp"
 #include "storage/storage_engine.h"
 #include "storage/virtual_column_utils.h"
 #include "types/json_value.h"
