+++ be/src/storage/primitive/predicate_parser.h
@@ -15,10 +15,13 @@
 #pragma once
 
 #include <cstdint>
+#include <vector>
 
+#include "common/column_id.h"
 #include "common/statusor.h"
 #include "storage/primitive/filter_condition.h"
 #include "storage/primitive/predicate_tree/predicate_tree_fwd.h"
+#include "types/type_info.h"
 
 namespace starrocks {
 
@@ -28,6 +31,11 @@ class SlotDescriptor;
 class RuntimeState;
 
 class ColumnPredicate;
+
+namespace predicate_parser_detail {
+ColumnPredicate* create_column_predicate(const TCondition& condition, TypeInfoPtr& type_info, ColumnId index);
+ColumnPredicate* create_column_predicate(const GeneralCondition& condition, TypeInfoPtr& type_info, ColumnId index);
+} // namespace predicate_parser_detail
 
 class PredicateParser {
 public:
@@ -48,6 +56,32 @@ public:
                                                       ExprContext* expr_ctx) const = 0;
 
     virtual uint32_t column_id(const SlotDescriptor& slot_desc) const = 0;
+};
+
+class ConnectorPredicateParser final : public PredicateParser {
+public:
+    explicit ConnectorPredicateParser(const std::vector<SlotDescriptor*>* slot_descriptors)
+            : _slot_desc(slot_descriptors) {}
+
+    bool can_pushdown(const ColumnPredicate* predicate) const override;
+
+    bool can_pushdown(const ConstPredicateNodePtr& pred_tree) const override;
+
+    bool can_pushdown(const SlotDescriptor* slot_desc) const override;
+
+    StatusOr<ColumnPredicate*> parse_thrift_cond(const TCondition& condition) const override;
+    StatusOr<ColumnPredicate*> parse_thrift_cond(const GeneralCondition& condition) const override;
+
+    StatusOr<ColumnPredicate*> parse_expr_ctx(const SlotDescriptor& slot_desc, RuntimeState*,
+                                              ExprContext* expr_ctx) const override;
+
+    uint32_t column_id(const SlotDescriptor& slot_desc) const override;
+
+private:
+    template <typename ConditionType>
+    ColumnPredicate* t_parse_thrift_cond(const ConditionType& condition) const;
+
+    const std::vector<SlotDescriptor*>* _slot_desc = nullptr;
 };
 
 } // namespace starrocks
