+++ be/src/storage/rowset/ordinal_page_index.cpp
@@ -36,11 +36,11 @@
 
 #include <memory>
 
+#include "cache/mem_cache/page_handle.h"
 #include "common/logging.h"
 #include "fs/fs.h"
 #include "runtime/exec_env.h"
 #include "storage/primitive/key_coder.h"
-#include "storage/rowset/page_handle.h"
 #include "storage/rowset/page_io.h"
 
 namespace starrocks {
