+++ be/src/storage/rowset/page_io.h
@@ -37,11 +37,11 @@
 #include <vector>
 
 #include "base/string/slice.h"
+#include "cache/mem_cache/page_handle.h"
 #include "common/logging.h"
 #include "common/status.h"
 #include "gen_cpp/segment.pb.h"
 #include "io/seekable_input_stream.h"
-#include "storage/rowset/page_handle.h"
 #include "storage/rowset/page_pointer.h"
 namespace starrocks {
