+++ be/src/storage/rowset/parsed_page.h
@@ -36,10 +36,10 @@
 
 #include <memory>
 
+#include "cache/mem_cache/page_handle_fwd.h"
 #include "storage/primitive/range.h"
 #include "storage/primitive/rowid_types.h" // ordinal_t
 #include "storage/rowset/page_decoder.h"
-#include "storage/rowset/page_handle_fwd.h"
 #include "storage/rowset/page_pointer.h"
 
 namespace starrocks {
