+++ be/src/storage/rowset/rowset_options.h
@@ -20,11 +20,11 @@
 
 #include "column/column_access_path.h"
 #include "column/global_dict/types.h"
+#include "compute_env/runtime_range_pruner.h"
 #include "storage/olap_common.h"
 #include "storage/options.h"
 #include "storage/primitive/predicate_tree/predicate_tree.hpp"
 #include "storage/runtime_filter_predicate.h"
-#include "storage/runtime_range_pruner.h"
 #include "storage/seek_range.h"
 #include "storage/tablet_schema.h"
