+++ be/src/storage/rowset/scalar_column_iterator.h
@@ -34,11 +34,11 @@
 
 #pragma once
 
+#include "cache/mem_cache/page_handle.h"
 #include "column/fixed_length_column.h"
 #include "storage/primitive/range.h"
 #include "storage/rowset/column_iterator.h"
 #include "storage/rowset/ordinal_page_index.h"
-#include "storage/rowset/page_handle.h"
 #include "storage/rowset/parsed_page.h"
 
 namespace starrocks {
