+++ be/src/storage/rowset/segment_options.h
@@ -22,12 +22,12 @@
 #include <vector>
 
 #include "column/global_dict/types.h"
+#include "compute_env/runtime_range_pruner.h"
 #include "storage/olap_common.h"
 #include "storage/options.h"
 #include "storage/primitive/disjunctive_predicates.h"
 #include "storage/primitive/predicate_tree/predicate_tree.hpp"
 #include "storage/runtime_filter_predicate.h"
-#include "storage/runtime_range_pruner.h"
 #include "storage/seek_range.h"
 
 namespace starrocks {
