+++ be/test/formats/parquet/file_reader_test.cpp
@@ -56,6 +56,7 @@
 #include "runtime/mem_tracker.h"
 #include "runtime/runtime_filter.h"
 #include "runtime/runtime_state.h"
+#include "storage/primitive/predicate_parser.h"
 #include "testutil/column_test_helper.h"
 #include "testutil/exprs_test_helper.h"
 #include "types/type_descriptor.h"
